Major adaptation to MyODBC architecture
Once we have engineered much of what a DES ODBC Driver should hold, this preliminary adaptation intends to make use of much of the preexistent code, so we could benefit from the enormous casuistry the MyODBC team considered and had to design accordingly. We adjusted the work done in order to fit into the MyODBC's architecture. The result is that, for instance, we are now using the descriptors as the ODBC standard intends without having added any additional code. The adaptation is not fully finished, but for now, the process of executing and retrieving results from a select query works appropriately.